data-ai-metrics-victoriametrics/pyproject.toml
[project]
name = "data-ai-metrics-victoriametrics"
version = "0.1.0"
description = "VictoriaMetrics-based Data Pipeline Observability with Polars, DuckDB, and Grafana"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "polars>=1.18.0",
    "duckdb>=1.1.3",
    "prometheus-client>=0.21.0",
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
    "openai>=1.59.5",
    "traceloop-sdk>=0.37.2",
    "faker>=33.1.0",
    "pyarrow>=18.1.0",
    "pandas>=2.0.0",
]
